load every mod script with import() (what can possibly go wrong?)
It seems that French-CC was the only mod which relied on `module:
false`, and now that this was fixed by @L-Sherry in
<L-Sherry/French-CC@d636b0a3b2d1ac5dc2cb0344e975bd373737d76c>,
there remains nothing preventing me from getting rid of this legacy
behavior. As an additional benefit now the modloader can collect
more exceptions thrown from the synchronous context of mod loading stage
files.