upgrade dependencies and lock some versions
Version of jszip is now locked at 3.5.0 because that's what we provide
with the repository, the npm package is used for the relevant type
declarations. Same goes for @types/semver, though that one is stuck at
semver 6.2.0 basically instead of our version which is semver 6.3.0.
That is because some new enum exported in 6.3.0 was tricky to write
definitions for, but fortunately the enum in question was the only
addition to the public API as far as I know, so the difference in
versions of the two packages doesn't really matter.